just: Check for file before removing in install-hooks
The `rm` command would fail if the hook did not exist.
First check for existence and then remove.

+++ justfile
@@ -153,7 +153,9 @@ install-hooks:
     #!/usr/bin/env bash
     set -e
     for hook in {{hooks}}; do
-        rm ".git/hooks/$hook"
+        if [ -f ".git/hooks/$hook" ]; then
+          rm ".git/hooks/$hook"
+        fi
         cp {{hook-script}} ".git/hooks/$hook"
         chmod +x ".git/hooks/$hook"
     done
